use correct casting for exp (maybe this fixes previous problems with MC19   scaling method)
svn-id: 332 @trunk

+++ PDSystemImpl/PDFullSpace/IpMc19TSymScalingMethod.cpp
@@ -130,7 +130,7 @@ namespace Ipopt
     // if that is the case
     Number sum=0.;
     for (Index i=0; i<n; i++) {
-      scaling_factors[i] = exp((R[i]+C[i])/2.);
+      scaling_factors[i] = exp((double)((R[i]+C[i])/2.));
       sum += scaling_factors[i];
     }
     if (!FiniteNumber(sum)) {
